requirements.txt
# requirements
celery==4.3.0
django==2.2.5
djangorestframework==3.10.1
django-cacheops==4.2
django-pg-bulk-update==3.0.0
emoji==0.5.4
funcy==1.13
gunicorn==19.9.0
keras==2.3.1
nltk==3.4.5
numpy==1.17.3
pathos==0.2.5
perceval==0.12.24
psycopg2==2.8.4
pymorphy2==0.8
pytz==2019.2
redis==3.3.8
sentry-sdk==0.10.2
six==1.12.0
sqlparse==0.3.0
statsd==3.3.0
tensorflow==2.0.0
typing==3.7.4